requirements.txt
flask==2.3.3
python-dotenv==1.0.0
google-generativeai==0.3.1
supabase==2.0.3
pandas==2.1.1
